requirements.txt
django==4.0.2
djangorestframework==3.13.1
pillow==9.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework-simplejwt==5.1.0